Add tests for the examples in the Ranges section of the spec.
I didn't make the range.order example into a test because I couldn't think of
a good way to handle it without changing the text of the example in a
negative way.

Add '&' to the list of characters extract_tests.py can un-escape and escaped
a few underscores in Modules.tex.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16689 3a8e244f-b0f2-452b-bcba-4c88e055c3ca