Reworked the implementation of queries in formal argument types so that they are handled more completely during function resolution and only marginally processed during normalization.
Fixed a bug in which the type of a field (with an omitted type) could
not be queried.  For example, this bug shows up in the following:

  class C { var x; }
  var c: C(int) = new C(2);
  f(c);
  def f(c: C(?t)) { }

Added a test to cover this case.  Before this fix, this was getting a
type-value error because the query was assumed to be of a value.
However, the type specification for a class with a field with an
omitted type requires a type.

Added a few more tests and updated the specification to discuss
implicit where clauses.  Implicit where clauses result when a formal
argument type is specified with some queried parts and other
non-queried parts.  For example

  def f(c: C(?t,real)) { ...

is translated to

  def f(c: C(?t,?tt)) where tt = real { ...

This has always worked in the implementation but was missing from the
spec.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16692 3a8e244f-b0f2-452b-bcba-4c88e055c3ca