Fixed reindexing for Block.  Adjust bounding box of the distribution (tricky part is avoiding over/underflow).  Added another argument to dsiCreatedReindexDist that is is the shape of the original domain.
Added a couple of tests that should produce error.

test_reindex5.chpl is failing for distString='block' because the .good
file needs to be different for Block (will work on trying to do that
with the test system).

test_composite2.chpl is failing.  Not quite sure what is the cause
(reindexing or other).  Investigating.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16695 3a8e244f-b0f2-452b-bcba-4c88e055c3ca