Added support to query the types of the fields of a tuple as in
  def f(x: (real, ?t))

Fixed a bug in which queries could not be parsed inside tuples and
reworked the parsing of formal array types to avoid a grammar
conflict.

Reworked the implementation of queries to handle the special case of
tuples and to also allow the compiler to determine the type of a
formal argument with a queried argument later in compilation (during
function resolution rather than normalization).  This latter change
will be important as we move towards better support for things like
policy classes where we may want to pass a generic type to a function
and use it as such.

Added a note about the implicit where clause added to a function with
a tuple formal argument where the type of at least one field is
queried.  In this case, an implicit where clause is added to the
function to ensure that the size of the actual tuple matches the size
of the formal tuple.

Removed an Aug '07 future regarding queries in tuples.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16698 3a8e244f-b0f2-452b-bcba-4c88e055c3ca