When creating a rank changed block distribution, make sure that the tuple that is used to slice the bounding box contains only scalars that are idxType or ranges over idxType.
The tuples that are used to slice the targetLocs array from block dsiRankChange
functions need to be tuples of int or range(int) values instead of idxType.

If a domain is rank change sliced with a scalar outside of the range for that
dimension, return an empty domain with the dimensions all set to 1..0.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16704 3a8e244f-b0f2-452b-bcba-4c88e055c3ca