Changed the algorithm for determining new bounding box for dsiCreateReindexDist after discussion with Steve and David.
NOTE: I used David's original addOverflowsType function (and added
subUnderflowsType), complete with conditional statements because I
felt this version was easier to reason about for someone not in on the
discussions, and so I wanted it documented (I put Steve's version with
conditional expressions in the comments)..  We still need to hammer
out the details of a more general version, and we can replace it at
that point if people find it offensive.

Fixed an output file due to change in error message (error conditions
checked in a different order now).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16706 3a8e244f-b0f2-452b-bcba-4c88e055c3ca