Completed a major rewrite of the Tuples chapter of the spec related to yesterday's spec review.  It could probably benefit from a second spec review meeting (though there are other chapters that should probably come first...).
Added spec tests for all examples.

Added more index entries.

There is a little implementation work to catch the implementation up
to the spec.  There are two futures in the spec tests related to this
work.  The two tasks are as follows: Fix tuple destructuring order in
the implementation and fix a bug related to grouping formal arguments
in a tuple in which it is impossible to specify a formal argument
type.

There are a few open issues in the spec.  Tuples of arrays are not
discussed in any depth and the future semantics of what it means to
create a tuple and what it means to pass a tuple to a function are not
mentioned anywhere.  Passing a tuple to a function should probably be
discussed in the Functions chapter.

This rewrite touches 5 other chapters in places.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16713 3a8e244f-b0f2-452b-bcba-4c88e055c3ca