Fix the multilocale distribution regressions for distString=block.
When rank change slicing a Block domain's bounding box, use the collapsed
dimension values plus the newRank sized index into targetLocDom to create the
index into the bigger targetLocDom.

When creating an index into the bounding box out of slice arguments, use the
bounding box low or high if the argument is out of range of the bounding box.

Added some comments to help explain what the code in the block rank change
functions is doing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16719 3a8e244f-b0f2-452b-bcba-4c88e055c3ca