Re-implemented tuple destructuring assignment.  Note that using tuples to swap values still works as I explain below.
Updated the spec and added a note about tuple creation, mentioning the
function call semantics and stating that the semantics of passing
tuples to functions is forthcoming.

In my changes to the Tuple section yesterday, I added a futurized test
of

  var x = 1, y = 2;
  (x, y) = (y, x);
  writeln((x, y));
    
and wrote that the correct output should be (2, 2) since we evaluate the
tuple destructuring assignment in order from left to right, and in tuple
component order.

This was wrong because the first thing that happens is that we evaluate
the entire right-hand side of the assignment statement.  This creates the
tuple (y, x).  The semantics are, as with a function call, to copy the
integral values.  So that tuple serves as a temporary, and we end up
swapping the values.  Tuple swap lives on!  But there are limitations...

Here is a program that really shows off the order of evaluation:

  var x = 1, y = 2;

  def L1() var { writeln("L1"); return x; }
  def L2() var { writeln("L2"); return y; }
 
  def R1() { writeln("R1"); return y; }
  def R2() { writeln("R2"); return x; }
 
  (L1(), L2()) = (R1(), R2());

Before this commit, the output was
  
  R1
  R2
  L2
  L1
  (2, 1)

which is right-first, then left, but the order of the left is
inverted.

With this commit, the output will be

  L1
  L2
  R1
  R2
  (2, 1)

which is left-first, then right.
  
Using the tuple notation to implement a swap doesn't always work
however.  This is because when we create a tuple of arrays, as with a
function call, we don't create temporary arrays (arrays are passed by
reference).  So the test:

  var A = [i in 1..4] i;
  writeln(A);
  (A(1..2), A(3..4)) = (A(3..4), A(1..2));
  writeln(A);

Before this commit, the output was

  1 2 3 4
  3 4 1 2

With this commit, the output will be

  1 2 3 4
  3 4 3 4

We used to insert a Chapel-level temporary (a temporary array), but we
no longer do.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16720 3a8e244f-b0f2-452b-bcba-4c88e055c3ca