Re-implemented tuple-grouped argument lists on functions so that the group can be specified with a type or a default value.  Individual arguments in the group cannot be specified with types or default values.  In particular, the following used to fail to parse:
  def f((x, y): (int, int)) { }

Removed the future on the test in the spec and updated the spec to
include the following:

  - A note on the implicit where clause that is attached to a function
    if any of its formal arguments are grouped using the tuple
    notation.

  - A list of restrictions on tuple-grouped arguments stating that:
    "Functions with tuple-grouped arguments may not be called using
    named-argument passing on the tuple-grouped arguments.  In
    addition, tuple-grouped arguments may not be specified
    individually with types or default values (only in aggregate).
    They may not be specified with any qualifier appearing before the
    group of arguments (or individual arguments) such as 'inout' or
    'type'.  They may not be followed by '...' to indicate that there
    are a variable number of them."

  - A new example that shows an equivalent function with an implicit
    where clause.

Added 6 test cases including one that combines a tuple type with
queried components resulting in a mess of implicit where clauses on
the function

  def f((x, y): (?u, real))

The implicit where clauses is something like

  def g(t: _tuple)
    where isTuple(t) && t.size == 2 && t.size == 2 && t(2).type == real

Note the duplication of both the check that t is a tuple and that
t.size is 2.  One is a result of the grouped arguments.  The other is
a result of the queried tuple type.  Since they are equivalent and
evaluated at compile-time, there is no loss.

Updated the .good for types/tuple/diten/overload_destructure_args
which exposes the name of the compiler-inserted temporary argument for
tuple-grouped arguments.  The name changed from _tuple_arg_tmp# to
chpl__tuple_arg_temp, but this shouldn't be showing up anyway.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16721 3a8e244f-b0f2-452b-bcba-4c88e055c3ca