Fixing an integer overflow issue in our memory query function
Due to a casting issue in our memory query function in C, we had
the potential to get overflow on systems with large memories and
32-bit longs that would result in us reporting a negative amount
of memory (which sometimes would show up in HPCC executions as
"can't take the log2() of a negative number".  The fix was a 
simple one -- cast the values being multiplied to be 64-bit ints
rather than the result of the multiplication of the two longs.
Should've done this years ago when first running into it.  The
move to stemp7 forced the issue as it's the first platform we
run regressions on to illustrate the problem.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16731 3a8e244f-b0f2-452b-bcba-4c88e055c3ca