Replaced support for using _ to omit expressions, variables, arguments, or indices when splitting a tuple (destructuring).  Instead of using _, now simply elide the component.  So
  Before this change        After this change
  ------------------        -----------------
  var (_,x) = (1,2);        var ( ,x) = (1,2);
  (_,x) = (1,2);            ( ,x) = (1,2);
  def f((_,x)) { }          def f(( ,x)) { }
  for (_,j) in I do         for ( ,j) in I do

Updated the spec.  Added a section on destructuring tuples in for
loops to the Tuples section and updated the section on for loops with
a forward reference.

Removed the restriction that variables cannot be simply named "_" and
updated the spec.

Updated regression tests.

Note that since _ is now a valid identifier, the following is still
legal

  var (_,x) =

and equivalent to

  var ( ,x) =

The following is not legal though

  var (_,x,_) =

since the variable _ is defined twice.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16734 3a8e244f-b0f2-452b-bcba-4c88e055c3ca