Simplifying Math module's use of extern calls
Thanks to Steve's bug fixes for external routines, the Math module has
been greatly simplified now.  In particular, extern routines whose
names match Chapel's no longer need to be wrapped in a Chapel
function.  Thus we can have an external definition of sin() rather
than a Chapel function that wraps the external definition of sin().
For 32-bit versions where we want to support the same overloaded name,
we still have to wrap it in order to (a) expose it using the name
sin() and (b) avoid putting sinf() and the like into the module's
namespace.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16736 3a8e244f-b0f2-452b-bcba-4c88e055c3ca