Adding HPL to the distribution robustness suite.
This is the same version as in release/examples with the arrays
distributed by Dist2D and the local block removed in schurComplement.

The problem size used for nightly testing is reduced to 40x40 to
reliably avoid timing out for the block distribution.  The bulk of the
time is spent performing remote gets, which is not unexpected for
block.

I have confirmed that this works for gasnet on our workstations as
well as on seal (XT system).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16750 3a8e244f-b0f2-452b-bcba-4c88e055c3ca