Removed support for using => on arrays that have already been declared.  It used to be restricted to new variable declarations. Conveniently, the spec was never updated to suggest that you could use the alias operator on existing arrays or array aliases.
Added initial support for alias-named-argument passing for
constructors so that an array field can be an alias to an array passed
to the constructor.

Test Changes and Additions:
--------------------------

Fixed arrays/diten/arrayAliasing and arrays/diten/arrayAliasing2D to
use the array alias operator only on new variable declarations.  My
other option was to make them fail with a syntax error, but this seems
better in preserving the functionality of these tests.

Added a test of using => on an array that has already been declared to
make sure the compiler issues an error.

Added a test of using => to alias an array passed to a function (not a
constructor) to make sure the compiler issues an error.

Module Changes:
--------------

Reworked the BlockDist, CyclicDist, and BlockCycDist to use
alias-named-argument passing instead of the alias operator to
construct class instances with aliases.

Removed the dsiCreateAlias method from the array classes for the
Default, GPUDist, and BlockDist distributions.  Removed the makeAlias
function that implemented => on arrays that have already been
declared.

Added a function chpl__getDomainFromArrayType that takes an array type
and returns the domain value associated with it.

Refactored the code that builds up the realms to use
alias-named-argument passing when setting up the Locales array field.
I had to change this code to use the default constructor in order to
support this functionality.  When we support alias-named-argument
passing to user-defined constructors, we can refactor this code to use
a user-defined constructor and remove the 'initialize' method that I
also added for the realm class.  Note that the only thing that the
'initialize' method does is to issue an error if a realm is created
too late during program execution.

Compiler Changes:
----------------

compiler/parser/chapel.ypp:

  Added support for using => when passing actuals to functions by name
  in order to support alias-named-argument passing for constructors.

compiler/include/build.h,
compiler/AST/build.cpp:

  Added functions to support parsing named actuals, aliases or not.

compiler/passes/scopeResolve.cpp:

  Added code to deal with alias-named-argument passing in two steps:

    1. Collect the names specified in alias-named-argument passing.

    2. For each field in a class that has one of these names, add a
    field with the same name prepended by "chpl__aliasField_" to allow
    for alias-named-argument passing.  Give this field a default value
    so that it does not have to be specified when constructing the
    class.  Each alias-named argument is parsed as two alias-named
    arguments, one for this new field to indicate that the array ought
    to alias.

compiler/include/flags.h,
compiler/AST/flags.cpp:

  Added a new flag "implicit alias field" used to mark fields in
  classes that were inserted by the compiler to support constructing
  classes with fields that are array aliases.

compiler/resolution/wrappers.cpp

  Added code to alias an array passed to a constructor if the implicit
  alias field is specified.  Since constructors are built up in their
  default wrapper, this is a natural place to handle this
  functionality.  To create the alias, the default wrapper gets the
  domain from the runtime array type via chpl__getDomainFromArrayType
  and reindexes via this domain.

compiler/passes/buildDefaultFunctions.cpp

  Add code to ignore the implicit alias fields when building
  compiler-inserted default methods and functions.

Future Work:
-----------

  * Support for aliasing array fields via user-defined constructors.
  * Support for aliasing array fields without specified types.
  * Document aliasing array fields in the spec.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16754 3a8e244f-b0f2-452b-bcba-4c88e055c3ca