Added the ability to index into 1-dimensional arrays with a 1-tuple of the base index type.  This applies to all array types.
For sparse and default array classes, the dsiAccess function only
needs to be specified to take an argument of type rank*idxType; for
opaque and associative array classes, the dsiAccess function only
needs to be specified to take an argument of type idxType.  I haven't
simplified these functions yet.

For the default sparse array class, I added an unfortunate conditional
to de-tuple a 1-tuple index when calling BinarySearch.  I'm not sure
whether this shows that we need a coercion or whether there is an
issue with the default sparse array class.  I left a comment here.

Updated the spec to mention indexing into arrays with 1-tuples.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16774 3a8e244f-b0f2-452b-bcba-4c88e055c3ca