Adding support for multiple mpirun flavors while removing hardcoded paths
This checkin permits the mpirun launcher to use multiple flavors of
mpirun (say, mpich1 vs. openmpi) while also removing the somewhat
hardcoded (and dependent on CHPL_HOME being set at execution time)
location of mpirun from the sources.  It does this by adding a util
script chplmpi.pl that returns which version of mpi you're wanting
to use as set by CHPL_MPI_VERSION or inferred (openmpi by default,
mpich1 if using armci).  It sets the location of mpirun in the
Makefile (eventually this should be over-ridable using an execution 
time environment variable as well, presumably).

Updated the printchplenv.pl script to print CHPL_MPI_VERSION if
mpi is part of the mix, and also to print the CHPL_TIMERS value.

Added the new script to the release tarball.

One svn:ignore that I missed on my last commit.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16790 3a8e244f-b0f2-452b-bcba-4c88e055c3ca