Refactored the Random module based on notes from this past Tuesday's code review.
Random.chpl:
-----------

Recommented and added a list of open issues.

Divided the class and module into public and private parts, prefixing
the private fields, methods, variables, and functions with the name of
the class/module followed by an underscore.

Created a new function getRandomStreamClockSeed to compute a default
seed for the RandomStream class based on the current time.  See
discussion below.

Added a compiler error to the fillRandom function if the element type
of the array is neither complex nor real.

Added a runtime warning if the RandomStream class is initialized with
an even seed.  Added a user-defined constructor.

Unified the fillRandom methods so there is just one piece of code.
The complex case is handled by making the iterator return a tuple of
random values in that case, and a single random value otherwise.

Reimplemented the parallel and serial iterators to take the array as
an argument rather than the number of elements.  The array is unused
except for getting a zero-based domain (see below) to capture the
shape of the iteration and also to look at the element type to
determine if the array is complex and requires pairs of random values.
Made the parallel iterator work for strided follower blocks and also
for multi-dimensional follower blocks.

DSIUtil.chpl:
------------

Added a computeZeroBasedDomain function that takes a domain and
returns a new domain with the same rank, index type, and shape, but
for which the indices in each dimension start at zero and have unit
stride.

tests/*:
-------

Updated tests.  These changes have to do with the removal of the
global variable SeedGenerator and replacements of
SeedGenerator.clockMS with the the function call
getRandomStreamClockSeed().  I didn't realize this variable was
actually used when we decided to look into simplifying this during the
code review.  This thus changes the interface.  Here are a few
alternatives (1 is implemented with this commit):

  1. Add a function getRandomStreamClockSeed() that returns an odd
     number based on a call to the function getCurrentTime.  This is
     the default seed initialization for creating a new RandomStream,
     or the user can pass this in explicitly.

  2. Revert to a SeedGenerator state with a variety of methods on it
     that can be used to generate seeds.  I think the disadvantages
     are having state when we don't need state.  Both functions and
     state impact the namespace.

  3. Do not expose this functionality.  If the user wants to
     initialize the seed in the default way, they can just not specify
     a seed.  The advantages here is not to provide a function that
     doesn't do all that much.  We still need to look into the issue
     of why we don't like even numbers.  If these are okay, then the
     user could just initialize to
       getCurrentTime(TimeUnits.microseconds):int(64)
     which is more explicit.

In addition, I updated the seed for some of these tests so that it is
odd.  It was relying on us adding one if it was even.  I left one test
that explicitly tests an even seed and updated its .good to capture
the warning we now generate.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16794 3a8e244f-b0f2-452b-bcba-4c88e055c3ca