Fix one more problem, this one found by valgrind testing.  It was possible for one thread to remove a begun task from the pool and free its descriptor, when another thread was still executing that task and thus using the descriptor.  Specifically, this could happen for tasks executed by CHPL_EXECUTE_TASKS_IN_LIST(), because such tasks were not necessarily at the head of the pool and so we didn't remove them from it (which would have been an expensive operation given the singly linked nature of the pool).  Prior to the tasks/threads ID change (r16711), we avoided this problem by not marking such tasks as "begun", preventing remove_begun_tasks_from_pool() from removing them from the pool.  But that was an obscure (and uncommented) solution.
The new solution is to use a doubly linked list for the pool, which
allows us to immediately remove a task from the pool as soon as we
decide to run it.  Doing so then allows us to put the responsibility for
freeing the task descriptor on the thread that runs that task, instead
of some other thread which happens to encounter the task later, after it
has worked its way to the front of the pool.  The cost, of course, is
another pointer in the descriptor and additional code to maintain the
doubly linked list.  Overall this seems a small price to pay, and may
even cost less than having to deal with having already-executed tasks in
the pool.  (If nothing else it reduces lock contention, since we now
have one fewer section of code protected by the threading_lock.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16797 3a8e244f-b0f2-452b-bcba-4c88e055c3ca