Miscellaneous improvements/changes to Random module
Recommented the standard Random module, quoting more of the comments
from the NPB benchmarks and noting the restrictions on the seed.

Replaced the warning, which was issued if a RandomStream object were
initialized with an even seed, with a halt, which is issued if a
RandomStream object is initialized with an even seed, a seed less than
or equal to zero, or a seed greater than or equal to 2**46.  Applied
the modulus operator to the time-based seed to ensure it is less than
2**46.

Loosened the element type check on fillRandom to also allow imag along
with real and complex.

Made parSafe method arguments parameters to avoid runtime overhead.

Added a serial/parallel iterator to the RandomStream class that is
invoked through the method iterate.  This method takes a domain, an
optional resultType type (default real), and the optional parSafe
parameter.  You can now do something like

  var rs = new RandomStream();
  forall r in rs.iterate(D) do ; // where D is a domain

  forall (a,r) in (A,rs.iterate(A.domain)) do
    a = r;

One proposal I have is to overload 'this' on the RandomStream class so
that indexing into rs with a domain will return the iterator, but
indexing into the stream with an integer would be equivalent to
getNth.  Arguably indexing into it with nothing could be equivalent to
getNext().  A method reset could be used to reset the stream.  Then
the above could be:

  forall r in rs(D) do ;

  A = rs(A.domain);


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16801 3a8e244f-b0f2-452b-bcba-4c88e055c3ca