Added a reindexing test that performs a reindex that will fail for block (passes for default and cyclic).
The prediff script will look at the compopts and copy the appropriate
.good file for the distribution to test_reindexing6.good.  For
example, when running with distString=block,
test_reindexing6.block.good will be copied.  If no such .good file
exists (as for cyclic), the default version will be copied.  This is a
generic script, and if it we seem to use this frequently, we should
use it as a directory level PREDIFF script.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16811 3a8e244f-b0f2-452b-bcba-4c88e055c3ca