Fixed the calculation of the "low" index in Cyclic.dsiCreateReindexDist.  This fixes test_strided_reindex1.chpl with distString=cyclic.
Add test_strided_reindex2.chpl test that does a reindex that changes stride
and doesn't work for cyclic. It was also failing for Block because of a bug
with named argument array aliases in constructors. The test
constructorArrayAliasReindex.chpl demonstrates the bug. The argument
used for the array alias in a constructor can't have a different stridable
parameter than the class is declared over, even though it would be reindexed
to match.

To work around that issue, I made a temp variable that is an alias declared
with a reindex to the right domain, and passed it to the constructor by alias
instead of passing the original version of the array.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16818 3a8e244f-b0f2-452b-bcba-4c88e055c3ca