The libatomics script is a hack and does not implement a general solution for determining the location of the libatomics library.
I don't see any immediate value in fixing this because:
(a) Its highly unlikely that the libatomics library will be made 
part of the Chapel distribution. 
(b) I am not certain about its very need in future, esp. if we
decide to move to GASNet atomics. 

The CHPL_STM=none implementation does not require this library. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@16830 3a8e244f-b0f2-452b-bcba-4c88e055c3ca