1. Implemented a 'fast followers' optimization (disable via    --no-fast-followers) that replaces the fragile 'aligned arrays    analysis' and fixes the recent performance slip of HPCC STREAM and    other basic programs.
2. Implemented a new, more readable mechanism for building up the AST
   after parsing but before normalization.

3. Fixed a bug in which copying a for loop would not update the break
   and continue labels in BlockStmt.

Fast Follower Details:
---------------------

To enable the 'fast follower' optimization for an array or domain,
simply define the following method:

  dsiStaticFastFollowerCheck(type leadType) param : bool

This method takes a type to determine if the fast follower can apply
given a leader of this type.  The type leadType is always a domain or
array class type.  For the Block and Cyclic distributions, I've
defined this method on the array class to return true if and only if
the lead type is the identical cyclic array type or the array's cyclic
domain type.

If this is not defined, nothing else has to be done and the array or
domain will not support the 'fast follower' optimization.

If this is defined, a dynamic check method should be defined for both
arrays and domains:

  dsiDynamicFastFollowCheck(lead: domain) : bool
  dsiDynamicFastFollowCheck(lead: []) : bool

For the Block and Cyclic distributions, I've defined this method to
return true if and only if the distribution of the block or cyclic
array or domain is the same as the distribution of the lead argument.

When these methods are defined, an extra param bool argument named
'fast' should be added to the follower.  This will be true if the
dynamic check passes, otherwise false.

This optimization works as follows: When 'parsing' a forall loop, we
generate code that generates two loops in a conditional, a fast
follower loop and a normal follower loop.  The conditional is guarded
by or'ing the result of calling the static checks on all of the
followers, and then and'ing the result of calling the dynamic checks
on all of the followers.  That is, if any follower has a fast version
that may be called, let's generate the code, but we're only going to
run the fast loop (all fast followers) if every follower that may be
called should be called.

Added a flag --[no-]fast-followers to disable/enable this
optimization.

Added a test to verify that this optimization fires.  To check this, I
added a config param in BlockDist that generates output whenever a
follower is invoked to distinguish between the fast and regular
followers.  Added a second test to ensure that the flag can turn off
this optimization.

Expr Build Details:
------------------

To build up the AST with this new mechanism, you can build expressions
directly using the new_Expr function, or you can inline these
arguments directly in BlockStmt::insertAtTail.  The intention is to
extend these to the other AST insertion/replace functions eventually,
if there is general buy-in to this mechanism.

The expected arguments are

  (const char* format, ...)

The format string should contain a Chapel statement or expression.
For example, given the declaration

  BlockStmt* block;

then the code

  block->insertAtTail("foo()");

is equivalent to

  block->insertAtTail(new CallExpr("foo"));

Symbols and expressions can be added to the newly created expressions
using %S and %E format flags.  For example, given the additional
declarations

  VarSymbol* tmp;
  CallExpr* call;

then the code

  block->insertAtTail("foo(%S)", tmp);
  block->insertAtTail("foo(%E)", call);

is equivalent to

  block->insertAtTail(new CallExpr("foo", tmp));
  block->insertAtTail(new CallExpr("foo", call));

Primitives can be defined by enclosing the name of the primitive in
apostrophes.  So the code

  block->insertAtTail("'move'(%S, new CallExpr("foo"))");

is equivalent to

  block->insertAtTail(new CallExpr(PRIM_MOVE, tmp, new CallExpr("foo")));

String literals are also supported by enclosing a string in
apostrophes.

Finally, block statements and type block statements
(BlockStmt::blockTag == BLOCK_TYPE) are supported via curly brackets
and semicolons.

As an example of the clean-up, the code

  leadBlock->insertAtTail(new CallExpr(PRIM_MOVE, leadIter, new CallExpr("_getIterator", new CallExpr("_toLeader", iter))));

becomes

  leadBlock->insertAtTail("'move'(%S, _getIterator(_toLeader(%S)))", leadIter, iter);


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16836 3a8e244f-b0f2-452b-bcba-4c88e055c3ca