Add knobs for on locale data parallelism for the Block distribution These knobs replace tasksPerLocale.
 * maxDataParallelism
 * limitDataParallelism
 * minDataParallelismSize

The knobs mirror the knobs added for default domains/arrays and
ranges.  The default values are set to the corresponding global values
for the default domains/arrays and ranges.

Added functions getMaxDataParallelism(), getLimitDataParallelism(),
and getMinDataParallelismSize() in DSIUtil.chpl to enable specifying
of default values of arguments to the Block constructor that have the
same name as the globals.

Changed test programs that referenced tasksPerLocale.  Updated .good
files as needed.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16871 3a8e244f-b0f2-452b-bcba-4c88e055c3ca