Fixing --gen-communicated-structures for homogenous tuples
When Steve changed homogenous tuples to use a static C array in the
generated C code, this broke our generated code that describes the
type/memory layout of communicated data structures because it still
treated it as a normal record (since that's how it's represented
within the compiler) yet there were no fields to obtain "offsetof()"
values from.

This commit fixes this issue, though in a fairly weak way.  It
essentially iterates over the homogenous tuple fields generating
"i*sizeof(fieldType)" offsets for the elements one by one.  This is no
worse than what we were doing before, yet requires O(n) work for an
n*type tuple.  That said, doing something smarter would be very
challenging given that the tuple could be a tuple of records in which
case we'd need to generate offsets and types for each of the fields
within that record (and it would also require bigger changes to the
PVM communication layer as well).

With any luck, this should be sufficient to get pvm testing up and
running again assuming that there haven't been other things falling
apart in the meantime.  Unfortunately, I only thought of this fix in
our recent planning meeting.

Added the --gen-communicated-structures flag to the OPT_TESTING_FLAGS
list in order to test this capability independently of PVM and keep it
in good shape over time.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16886 3a8e244f-b0f2-452b-bcba-4c88e055c3ca