Removing sub_test and replacing it with sub_test.py
Keeping both sub_test and sub_test.py around has not proven to be
useful in that the old sub_test no longer worked on cygwin and
we weren't necessarily keeping it up-to-date.  So, removed sub_test
and renamed sub_test.py to sub_test.

Simplified the logic in start_test that was trying to determine
whether or not it should use the python version or not -- now
it always uses the python version.

Made the sub_test scripts in testing directories invoke sub_test
rather than sourcing it now that it's no longer a csh script (this
seemed to still work, though I was a little surprised that it did).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16887 3a8e244f-b0f2-452b-bcba-4c88e055c3ca