This change breaks up chpl_main into chpl_main and chpl_user_main.
This change breaks up chpl_main into user and Chapel code. Before, the
runtime would call chpl_main, which would initialize the program, run the
program, and gather memory statistics. Now, chpl_main intializes the
program and gathers memory statistics, but it calls chpl_user_main to
execute the user code. One side effect is the call stack is now one item
deeper, and so to avoid chpl_user_main showing up on dumps, we define
chpl_main to have FLAG_TEMP such that when we call the stack, chpl_main
is ignored. To allow this to work with CHPL_DEVELOPER, in insertLineNumbers,
we make sure we're not in chpl_main. (Note: instead of using strcmp to
check this, we should add FLAG_EXPORT -- this method could rid all direct
string fn->name comparisons.)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16888 3a8e244f-b0f2-452b-bcba-4c88e055c3ca