Changing linux to linux32
Since linux64 has become the standard, calling 32-bit linux simply
'linux' seemed increasingly inappropriate, so this renames it to
linux32.

Just to capture something that was pointed out on email:

Q: Why do we distinguish between linux32 and linux64 anyway?

A: Historically it's been so that a single installation could
   have builds for both architectures.  We used to use both
   regularly so this is helpful.  Going forward it may be less
   so, so at some point in the future, we may want to simply
   use 'linux' for both.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16896 3a8e244f-b0f2-452b-bcba-4c88e055c3ca