Adding multicore/multitasking hello world example
In walking people through installation of the Chapel release on their
desktops, the first question I usually get asked is how they can get a
multicore program up and running in short order to make sure it's
working, which led me to add this variation on hello world that uses
multiple tasks on the current locale.

Updated the documentation/Makefiles to describe this test.

On Steve's suggestion, removed the config const from the multilocale
hello world program -- it's an unnecessary detail that's better
introduced in the structured version.

Updated documentation for this directory



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16899 3a8e244f-b0f2-452b-bcba-4c88e055c3ca