This change adds a new pass to localize all global constants.
This change adds a new pass to localize all global constants. The motivation
is to reduce the number of loads (in loops) to optimize performance. It is
defaulted to be off (tested both on and off), though it can be enabled with
a new flag, fNoGlobalConstOpt, enabled with localize-global-consts. We don't
do this in initialization functions (to prevent a local copy of a global
constant getting thrown out as soon as the initialization finishes). This
has been tested on XMT to show increased performance, though only
regression analysis has been done on other systems.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16901 3a8e244f-b0f2-452b-bcba-4c88e055c3ca