Add spec tests for examples in section "Parallelism and Synchronization"
Added a "chapelprediff" environment in chapel_testing.tex and extract_tests.py
to create executable prediff files for tests. Added "$\mbox{{\bf \$}}$" to
the list of strings that get substituted out of extracted tests. It gets
translated to "$".

test_serialStmt2.chpl fails because it uses a serial iterator in a forall loop.

test_forallFilter.chpl looks like it was meant to filter every other character
out of a string, but is currently using an array of ints instead.

Tests involving atomic sections currently use the --no-warnings flag to avoid
the "atomics not implemented" warning.

Fixed bugs in some examples: a recursive function needed an explicit return
type, a recursive binary tree walking function was calling an undefined
function instead of itself.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16906 3a8e244f-b0f2-452b-bcba-4c88e055c3ca